+ Porting Progress 3

+++ CustomCurrency.cs
@@ -11,7 +11,7 @@ namespace AAMod
 		{
 		}
 
-		public override void GetPriceText(string[] lines, ref int currentLine, int price)
+		public override void GetPriceText(string[] lines, ref int currentLine, long price)
 		{
 			Color goblinColor = Color.ForestGreen * (Main.mouseTextColor / 255f);
 			lines[currentLine++] = string.Format("[c/{0:X2}{1:X2}{2:X2}:{3} {4} {5}]", new object[]
