+++ Items/Currency/MartianCredit.cs
@@ -37,7 +37,7 @@ namespace AAMod.Items.Currency
         {
         }
 
-        public override void GetPriceText(string[] lines, ref int currentLine, int price)
+        public override void GetPriceText(string[] lines, ref int currentLine, long price)
         {
             Color color2 = color * (Main.mouseTextColor / 255f);
             lines[currentLine++] = string.Format("[c/{0:X2}{1:X2}{2:X2}:{3} {4} {5}]", new object[]
